moved ros and bootstrap to algo submodule
