+++ wqio/__init__.py
@@ -1,5 +1,5 @@
-from .core import *
 from . import utils
+from .core import *
 
 from .testing import NoseWrapper
 test = NoseWrapper().test
