+++ wqio/algo/__init__.py
@@ -1,0 +1,2 @@
+from . import ros
+from . import bootstrap
