+++ wqio/utils/__init__.py
@@ -2,12 +2,10 @@ from matplotlib.scale import register_scale
 
 from .misc import *
 from .exceptions import DataError
-from . import bootstrap
-from . import ros
 from .probscale import ProbScale
 from . import figutils
 
-from ..testing import NoseWrapper
+from wqio.testing import NoseWrapper
 test = NoseWrapper().test
 
 register_scale(ProbScale)
